Added user models and reorganized configs (#40)
