+++ backend/src/config.py
@@ -1,0 +1,2 @@
+# config.py
+DATABASE_URL = "postgresql://ct:password@localhost:5433/ct"
