+++ backend/src/models/base.py
@@ -1,0 +1,4 @@
+# base.py
+from sqlalchemy.orm import declarative_base
+
+Base = declarative_base()
